requirements.txt
required
python 3
django 2.0
djangorestframework
